154 separate arg handling from program class
